restore user

+++ content/moduls.php
@@ -4,13 +4,12 @@ $id_user = isset($_SESSION['ID_USER']) ? $_SESSION['ID_USER'] : '';
 $id_role = isset($_SESSION['ID_ROLE']) ? $_SESSION['ID_ROLE'] : '';
 
 $rowStudent = mysqli_fetch_assoc(mysqli_query($config, "SELECT * FROM students WHERE id='$id_user'"));
-$id_major = $rowStudent['id_major'];
+$id_major = isset($rowStudent['id_major']) ? $rowStudent['id_major'] : '';
 
-if ($id_role == 2) {
+if ($id_role == 7) {
     $where = "WHERE moduls.id_major='$id_major'";
-} elseif ($id_role == 1) {
+} elseif ($id_role == 5) {
     $where = "WHERE moduls.id_instructor='$id_user'";
-   
 }
 $query = mysqli_query($config, "SELECT majors.name as major_name, instructors.name as instructor_name, moduls.* FROM moduls LEFT JOIN majors ON majors.id = moduls.id_major LEFT JOIN instructors ON instructors.id = moduls.id_instructor $where ORDER BY moduls.id DESC");
 $row = mysqli_fetch_all($query, MYSQLI_ASSOC);
@@ -22,7 +21,7 @@ $row = mysqli_fetch_all($query, MYSQLI_ASSOC);
         <div class="card">
             <div class="card-body">
                 <h5 class="card-title">Data Modul</h5>
-                <?php if ($id_role == 1): ?>
+                <?php if (canAddModul(5)): ?>
                     <div class="mb-3" align="right">
                         <a href="?page=tambah-modul" class="btn btn-primary">Add Modul</a>
                     </div>
@@ -50,8 +49,10 @@ $row = mysqli_fetch_all($query, MYSQLI_ASSOC);
                                     <td><?php echo $data['instructor_name'] ?></td>
                                     <td><?php echo $data['major_name'] ?></td>
                                     <td>
-                                        <a href="?page=tambah-modul&edit=<?php echo $data['id'] ?>" class="btn btn-success btn-sm">Edit</a>
-                                        <a onclick="return confirm ('Are you sure?')" href="?page=tambah-modul&delete=<?php echo $data['id'] ?>" class="btn btn-warning btn-sm">Delete</a>
+                                        <?php if ($id_role == 1): ?>
+                                            <a href="?page=tambah-modul&edit=<?php echo $data['id'] ?>" class="btn btn-success btn-sm">Edit</a>
+                                            <a onclick="return confirm ('Are you sure?')" href="?page=tambah-modul&delete=<?php echo $data['id'] ?>" class="btn btn-warning btn-sm">Delete</a>
+                                        <?php endif ?>
                                     </td>
                                 </tr>
                             <?php endforeach; ?>
